requirements.txt
# tensorflow 2.15.0 already included in base image (tensorflow/tensorflow:2.15.0-gpu)
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
pillow==10.2.0
kaggle==1.6.6
jupyterlab==4.1.5
ipywidgets==8.1.2
ipykernel==6.29.3
tqdm==4.66.2
